♻️ DR6HVN integrate: decompose queue command
