+++ .agentplane/tasks/202605282111-DR6HVN/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605282111-DR6HVN`
+Title: Integrate queue command decomposition
+Canonical task record: `.agentplane/tasks/202605282111-DR6HVN/README.md`
+
+## Summary
+
+Integrate queue command decomposition
+
+Decompose packages/agentplane/src/commands/integrate-queue.command.ts into focused queue helper modules without changing branch_pr merge-lane behavior, hosted check waiting, stale lane recovery, or queue state transitions. Reduce the command file below the hotspot warning threshold when safe.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/integrate-queue.command.ts into focused queue helper modules without changing branch_pr merge-lane behavior, hosted check waiting, stale lane recovery, or queue state transitions. Reduce the command file below the hotspot warning threshold when safe.
+- Out of scope: unrelated refactors not required for "Integrate queue command decomposition".
+
+## Verification
+
+- State: ok
+- Note: Integrate queue command decomposition verified locally.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T21:12:29.916Z
+- Branch: task/202605282111-DR6HVN/integrate-queue-command-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
