+++ .agentplane/tasks/202605282111-DR6HVN/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T21:12:29.916Z
+
+## Task
+
+- Task: `202605282111-DR6HVN`
+- Title: Integrate queue command decomposition
+- Status: DOING
+- Branch: `task/202605282111-DR6HVN/integrate-queue-command-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282111-DR6HVN/README.md`
+
+## Verification
+
+- State: ok
+- Note: Integrate queue command decomposition verified locally.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T21:12:29.916Z
+- Branch: task/202605282111-DR6HVN/integrate-queue-command-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
